// src/types/flightType.ts
export interface dataTypes {
    airline: string;
    arrival_airport: string;
    arrival_city: string;
    arrival_country: string;
    arrival_country_popular: string;
    class: string;
    departure_airport: string;
    departure_city: string;
    departure_country: string;
    duration: number;
    flight_number: string;
    price: number;
    seats_available: number;
}